dug into the data to make sure I pulled out what I really wanted
