+++ src/components/App/App.jsx
@@ -1,6 +1,6 @@
 import { Routes, Route } from 'react-router-dom';
-import TeaSubscriptionContainer from './TeaSubscriptionContainer';
-import TeaSubscriptionDetails from './TeaSubscriptionDetails';
+import TeaSubscriptionContainer from '../TeaSubscriptionContainer/TeaSubscriptionContainer';
+// import TeaSubscriptionDetails from '../TeaSubscriptionDetails/TeaSubscriptionDetails';
 
 import './App.css'
 
@@ -10,8 +10,8 @@ function App() {
       <h1>🫖 The Koala Tea Control Admin View 🫖</h1>
 
       <Routes>
-        < Route path="/subscriptions" element={<TeaSubscriptionContainer />} />
-        < Route path="/subscriptions/:id" element={<TeaSubscriptionDetails/>} />
+        < Route path="/" element={<TeaSubscriptionContainer />} />
+        {/* < Route path="/subscriptions/:id" element={<TeaSubscriptionDetails/>} /> */}
       </Routes>
     </div>
   )
