+++ src/components/TeaSubscriptionCard/TeaSubscriptionCard.jsx
@@ -1,16 +1,17 @@
 import React from "react";
 
 const TeaSubscriptionCard = ({ subscription, toggleStatus }) => {
-    const { id, customer_name, tea_type, frequency, status } = subscription;
+// console.log(subscription.attributes)
+    const teaSub = subscription.attributes
 
     return (
         <div className="tea-subscription-card">
-            <h3>Customer: {customer_name} - Tea: {tea_type}</h3>
-            <p>Order delivery {frequency}</p>
-            <p>Subscription status: {status}</p>
+            <h3>Customer: {teaSub.customer_name} - Tea: {teaSub.tea_type}</h3>
+            <p>Order delivery: {teaSub.frequency}</p>
+            <p>Subscription status: {teaSub.status}</p>
 
-            <button onClick={() => toggleStatus(id, status)}>
-                {status === "active" ? "Deactivate Subscription" : "Reactivate Subscription"}
+            <button onClick={() => toggleStatus(id, teaSub.status)}>
+                {teaSub.status === "active" ? "Deactivate Subscription" : "Reactivate Subscription"}
             </button>
         </div>
     )
